UI Update: Added in Search Feature + other  minor misc UI updates
