+++ frontend/second-bite-vite-frontend/src/pages/MainPage.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import PrimaryHeader from '../header/PrimaryHeader';
 import ConsumerLocation from '../components/main/LocationSearch/ConsumerLocation';
-import FilterSort from '../components/main/FilterSort/FilterSort';
+import Specifiers from '../components/main/FilterSort/Specifiers';
 import RestaurantTiles from '../components/main/RestaurantTiles/RestaurantTiles';
 
 const MainPage = () => {
@@ -11,7 +11,7 @@ const MainPage = () => {
             <PrimaryHeader />
             <main>
                 <ConsumerLocation />
-                <FilterSort />
+                <Specifiers />
                 <RestaurantTiles />
             </main>
         </>
